chore: fix secondary-sidebar
