+++ frontend/pages/src/common/components/layouts/admin-panel/nav-user.tsx
@@ -4,15 +4,15 @@ import { Bell, ChevronsUpDown, LogOut, Settings2, UserIcon } from "lucide-react"
 
 import { useLogout, useProfilePictureUrl, useUser } from "@auth"
 import { Avatar } from "@incmix/ui"
-import { DropdownMenu } from "@radix-ui/themes"
-import { Link } from "@tanstack/react-router"
-import { useTranslation } from "react-i18next"
 import {
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
   useSidebar,
-} from "./sidebar"
+} from "@incmix/ui/sidebar"
+import { DropdownMenu } from "@radix-ui/themes"
+import { Link } from "@tanstack/react-router"
+import { useTranslation } from "react-i18next"
 
 export function NavUser({
   userId,
