+++ frontend/pages/src/common/components/layouts/admin-panel/org-switcher.tsx
@@ -4,15 +4,15 @@ import { ChevronsUpDown } from "lucide-react"
 import * as React from "react"
 
 import { useOrganizationStore } from "@incmix/store"
-import { useOrganizations } from "@orgs/utils"
-import { DropdownMenu } from "@radix-ui/themes"
-import { useTranslation } from "react-i18next"
 import {
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
   useSidebar,
-} from "./sidebar"
+} from "@incmix/ui/sidebar"
+import { useOrganizations } from "@orgs/utils"
+import { DropdownMenu } from "@radix-ui/themes"
+import { useTranslation } from "react-i18next"
 
 export function OrgSwitcher() {
   const { isMobile } = useSidebar()
