+++ frontend/ui/package.json
@@ -43,7 +43,11 @@
     "./omit": "./src/utils/omit.ts",
     "./preset": "./src/preset.ts",
     "./scroll-area": "./src/components/scroll-area.ts",
+    "./secondary-sidebars": "./src/components/secondary-sidebars/index.tsx",
+    "./secondary-sidebars/file-folder": "./src/components/secondary-sidebars/file-folder/index.tsx",
     "./sheet": "./src/components/sheet.tsx",
+    "./sidebar": "./src/components/sidebar/index.ts",
+    "./slider": "./src/components/slider.ts",
     "./slot": "./src/primitives/slot.ts",
     "./stepper": "./src/components/stepper/index.ts",
     "./split-button": "./src/components/button/split-button.tsx",
