+++ frontend/ui/src/components/scroll-area.ts
@@ -1,1 +1,1 @@
-export { ScrollArea, type ScrollAreaProps } from "@radix-ui/themes"
+export { ScrollArea } from "@radix-ui/themes"
