+++ frontend/ui/src/components/scroll-area.tsx
@@ -1,0 +1,1 @@
+export { ScrollArea, type ScrollAreaProps } from "@radix-ui/themes"
