+++ frontend/ui/src/components/secondary-sidebars/file-folder/icons/folder-open.tsx
