+++ frontend/ui/src/components/secondary-sidebars/file-folder/icons/types.ts
