+++ frontend/ui/src/components/sidebar/sidebar.tsx
@@ -1,335 +1,928 @@
+"use client"
+import * as React from "react"
+
+import { Slot } from "@radix-ui/react-slot"
+import { type VariantProps, cva } from "@utils/cva"
+import { ChevronsLeft, PanelLeft } from "lucide-react"
+
+import { useIsMobile } from "@hooks/use-mobile"
+import { cn } from "@utils/cn"
+import type { Button } from "../button/button"
+import { IconButton } from "../button/icon-button"
+import { Input } from "../form/input"
+import { Separator } from "../separator"
+import { Sheet, SheetContent, SheetDescription, SheetTitle } from "../sheet"
+import { Skeleton } from "../skeleton"
 import {
-  ChevronDownIcon,
-  PinLeftIcon,
-  PinRightIcon,
-} from "@radix-ui/react-icons"
-import { Box, DropdownMenu, Flex, Text } from "@radix-ui/themes"
-import { Link as RouterLink } from "@tanstack/react-router"
-import React, {
-  type MouseEventHandler,
-  useRef,
-  createContext,
-  useContext,
-} from "react"
-import { ReactiveButton } from "../button/reactive-button"
-import { useSidebarStore } from "./store"
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "../tooltip"
 
-export type SidebarState = "extended" | "minified" | "closed"
+const SIDEBAR_COOKIE_NAME = "sidebar:state"
+const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
+const SIDEBAR_WIDTH = "16rem"
+const SIDEBAR_WIDTH_MOBILE = "18rem"
+const SIDEBAR_WIDTH_ICON = "4rem"
+const SIDEBAR_KEYBOARD_SHORTCUT = "b"
 
-type SidebarContextType = {
-  minified: boolean
-  state: SidebarState
-  currentlyExpanded: () => boolean
+type SidebarContext = {
+  // Primary sidebar
+  state: "expanded" | "collapsed"
+  open: boolean
+  setOpen: (open: boolean) => void
+
+  // Secondary sidebar
+  secondaryOpen: boolean
+  setSecondaryOpen: (open: boolean) => void
+
+  // Common
+  isMobile: boolean
+  openMobile: boolean
+  setOpenMobile: (open: boolean) => void
+  toggleSidebar: () => void
+  toggleSecondarySidebar: () => void
 }
 
-const SidebarContext = createContext<SidebarContextType | null>(null)
+const SidebarContext = React.createContext<SidebarContext | null>(null)
 
-export function useSidebar() {
-  const context = useContext(SidebarContext)
+function useSidebar() {
+  const context = React.useContext(SidebarContext)
   if (!context) {
-    throw new Error("Sidebar components must be used within a Sidebar")
+    throw new Error("useSidebar must be used within a SidebarProvider.")
   }
+
   return context
 }
 
-export type SidebarProps = {
-  minified?: boolean
-  state?: SidebarState
-  children?: React.ReactNode
-  className?: string
-  name?: string
-  onMinifyChange?: (minified: boolean) => void
-}
+const SidebarProvider = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div"> & {
+    // Primary sidebar
+    defaultOpen?: boolean
+    open?: boolean
+    onOpenChange?: (open: boolean) => void
 
-const MINIFIED_WIDTH = "w-14"
-
-export function Sidebar({
-  minified = false,
-  state = "extended",
-  children,
-  className = "",
-  name,
-  onMinifyChange,
-}: SidebarProps) {
-  const sidebarWidth = minified
-    ? MINIFIED_WIDTH
-    : state === "extended"
-      ? "w-sidebar"
-      : "w-0"
-
-  const currentlyExpanded = () => state === "extended" && !minified
-
-  const contextValue = {
-    minified,
-    state,
-    currentlyExpanded,
+    // Secondary sidebar
+    defaultSecondaryOpen?: boolean
+    secondaryOpen?: boolean
+    onSecondaryOpenChange?: (open: boolean) => void
   }
+>(
+  (
+    {
+      defaultOpen = true,
+      open: openProp,
+      onOpenChange: setOpenProp,
+      className,
+      style,
+      children,
+      ...props
+    },
+    ref
+  ) => {
+    const {
+      defaultSecondaryOpen = true,
+      secondaryOpen: secondaryOpenProp,
+      onSecondaryOpenChange: setSecondaryOpenProp,
+    } = props
+    const isMobile = useIsMobile()
+    const [openMobile, setOpenMobile] = React.useState(false)
+    const [_secondaryOpen, _setSecondaryOpen] =
+      React.useState(defaultSecondaryOpen)
+    const secondaryOpen = secondaryOpenProp ?? _secondaryOpen
 
-  return (
-    <SidebarContext.Provider value={contextValue}>
-      <Box
-        asChild
-        className={`h-full select-none overflow-hidden bg-gray-3 transition-[width] duration-300 ease-in-out ${sidebarWidth} ${className}`}
-      >
-        <aside className="border-gray-6 border-r bg-gray-3">
-          <Box className="h-full w-sidebar bg-gray-3">
-            {name && (
-              <Flex
-                align="center"
-                justify="center"
-                px="4"
-                py="3"
-                className={`border-gray-6 border-b transition-opacity duration-300 ${
-                  minified ? "opacity-0" : "opacity-100"
-                }`}
-              >
-                <Text size="3" weight="medium">
-                  {name}
-                </Text>
-              </Flex>
-            )}
-            {children}
-            {onMinifyChange && (
-              <Box
-                className={`fixed bottom-4 transition-all duration-300 ease-in-out ${
-                  minified ? "translate-x-5" : "translate-x-[13.5rem]"
-                }`}
-              >
-                {
-                  <ReactiveButton
-                    variant="ghost"
-                    onClick={() => onMinifyChange(!minified)}
-                    className="p-2"
-                  >
-                    {minified ? <PinRightIcon /> : <PinLeftIcon />}
-                  </ReactiveButton>
-                }
-              </Box>
-            )}
-          </Box>
-        </aside>
-      </Box>
-    </SidebarContext.Provider>
-  )
-}
+    // This is the internal state of the sidebar.
+    // We use openProp and setOpenProp for control from outside the component.
+    const [_open, _setOpen] = React.useState(defaultOpen)
+    const open = openProp ?? _open
+    const setOpen = React.useCallback(
+      (value: boolean | ((value: boolean) => boolean)) => {
+        const openState = typeof value === "function" ? value(open) : value
+        if (setOpenProp) {
+          setOpenProp(openState)
+        } else {
+          _setOpen(openState)
+        }
 
-type SidebarItemBaseProps = {
-  icon: React.ReactNode
-  label: string
-  badge?: number
-}
+        // This sets the cookie to keep the sidebar state.
+        document.cookie = `${SIDEBAR_COOKIE_NAME}=${openState}; path=/; max-age=${SIDEBAR_COOKIE_MAX_AGE}`
+      },
+      [setOpenProp, open]
+    )
+    // Helper to toggle the sidebar.
+    const toggleSidebar = React.useCallback(() => {
+      setOpen((open) => !open)
+    }, [setOpen])
 
-type SidebarItemWithChildrenProps = SidebarItemBaseProps & {
-  children: React.ReactNode
-  to?: never
-  onClick?: never
-}
-
-type SidebarItemWithToProps = SidebarItemBaseProps & {
-  to: string
-  children?: never
-  onClick?: never
-}
-
-type SidebarItemWithOnClickProps = SidebarItemBaseProps & {
-  onClick: () => void
-  to?: never
-  children?: never
-}
-
-export type SidebarItemProps =
-  | SidebarItemWithChildrenProps
-  | SidebarItemWithToProps
-  | SidebarItemWithOnClickProps
-
-export function SidebarItem({
-  icon,
-  label,
-  to,
-  badge,
-  children,
-  onClick,
-}: SidebarItemProps) {
-  const { minified } = useSidebar()
-  const { isItemExpanded, toggleItem } = useSidebarStore()
-  const contentRef = useRef<HTMLDivElement>(null)
-
-  const toggleExpand: MouseEventHandler<HTMLButtonElement> = (e) => {
-    if (children && !e.defaultPrevented) {
-      e.preventDefault()
-      if (!minified) {
-        toggleItem(to || label)
-      }
-    }
-  }
-
-  const renderIcon = () => (
-    <Box className="flex h-6 w-6 items-center justify-center">{icon}</Box>
-  )
-
-  const renderLabel = () => (
-    <Text ml="3" className="flex-grow">
-      {label}
-    </Text>
-  )
-
-  const renderBadge = () =>
-    badge && (
-      <Box className="ml-auto rounded-full bg-blue-9 px-2 py-1 font-semibold text-white text-xs">
-        {badge}
-      </Box>
+    const setSecondaryOpen = React.useCallback(
+      (value: boolean | ((prev: boolean) => boolean)) => {
+        const newValue =
+          typeof value === "function" ? value(secondaryOpen) : value
+        setSecondaryOpenProp?.(newValue) || _setSecondaryOpen(newValue)
+      },
+      [setSecondaryOpenProp, secondaryOpen]
     )
 
-  const renderExpandIcon = () =>
-    children && (
-      <ChevronDownIcon
-        color="gray"
-        className={`ml-2 h-4 w-4 transition-transform duration-200 ${
-          isItemExpanded(to || label) ? "rotate-180" : ""
-        }`}
+    const toggleSecondarySidebar = React.useCallback(() => {
+      setSecondaryOpen((prev) => !prev)
+    }, [setSecondaryOpen])
+
+    // Adds a keyboard shortcut to toggle the sidebar.
+    React.useEffect(() => {
+      const handleKeyDown = (event: KeyboardEvent) => {
+        if (
+          event.key === SIDEBAR_KEYBOARD_SHORTCUT &&
+          (event.metaKey || event.ctrlKey)
+        ) {
+          event.preventDefault()
+          toggleSidebar()
+        }
+      }
+
+      window.addEventListener("keydown", handleKeyDown)
+      return () => window.removeEventListener("keydown", handleKeyDown)
+    }, [toggleSidebar])
+
+    // We add a state so that we can do data-state="expanded" or "collapsed".
+    // This makes it easier to style the sidebar with Tailwind classes.
+    const state = open ? "expanded" : "collapsed"
+
+    const contextValue = React.useMemo<SidebarContext>(
+      () => ({
+        state,
+        open,
+        setOpen,
+        isMobile,
+        openMobile,
+        setOpenMobile,
+        toggleSidebar,
+        setSecondaryOpen,
+        secondaryOpen,
+        toggleSecondarySidebar,
+      }),
+      [
+        state,
+        open,
+        setOpen,
+        isMobile,
+        openMobile,
+        toggleSidebar,
+        setSecondaryOpen,
+        secondaryOpen,
+        toggleSecondarySidebar,
+      ]
+    )
+
+    return (
+      <SidebarContext.Provider value={contextValue}>
+        <TooltipProvider delayDuration={0}>
+          <div
+            style={
+              {
+                "--sidebar-width": SIDEBAR_WIDTH,
+                "--sidebar-width-icon": SIDEBAR_WIDTH_ICON,
+                ...style,
+              } as React.CSSProperties
+            }
+            className={cn(
+              "group/sidebar-wrapper flex min-h-svh w-full has-[[data-variant=inset]]:bg-sidebar",
+              className
+            )}
+            ref={ref}
+            {...props}
+          >
+            {children}
+          </div>
+        </TooltipProvider>
+      </SidebarContext.Provider>
+    )
+  }
+)
+SidebarProvider.displayName = "SidebarProvider"
+
+const Sidebar = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div"> & {
+    side?: "left" | "right"
+    variant?: "sidebar" | "floating" | "inset"
+    collapsible?: "offcanvas" | "icon" | "none"
+    isDefaultMobile?: boolean
+  }
+>(
+  (
+    {
+      side = "left",
+      variant = "sidebar",
+      collapsible = "offcanvas",
+      className,
+      children,
+      isDefaultMobile = true,
+      ...props
+    },
+    ref
+  ) => {
+    const { isMobile, state, openMobile, setOpenMobile } = useSidebar()
+    const SideBarTrigger = (
+      <SidebarTrigger
+        icon={
+          <ChevronsLeft className="stroke-[hsl(var(--sidebar-background))]" />
+        }
+        className="-right-5 fixed top-10 left-[calc(var(--sidebar-width)_-_20px)] z-50 h-10 w-10 rounded-full border border-[hsl(var(--sidebar-trigger-border))] border-solid bg-[hsl(var(--sidebar-trigger-background))] transition-all duration-200 ease-linear group-data-[collapsible=icon]:left-[calc(var(--sidebar-width-icon)_-_20px)] group-data-[collapsible=icon]:rotate-180"
       />
     )
 
-  const renderChildren = () =>
-    !minified &&
-    children && (
-      <Box
-        ref={contentRef}
-        className={`overflow-hidden transition-all duration-300 ease-in-out ${
-          isItemExpanded(to || label) ? "max-h-96" : "max-h-0"
-        }`}
-      >
-        <Flex direction="column" gap="1" ml="6" mt="2" mb="2">
-          {React.Children.map(children, (child) => {
-            if (React.isValidElement(child)) {
-              const { label, to } = child.props
-              return <SidebarSubItem key={to} label={label} to={to} />
+    if (collapsible === "none") {
+      return (
+        <div
+          className={cn(
+            "flex h-full w-[--sidebar-width] flex-col bg-sidebar text-sidebar-foreground",
+            className
+          )}
+          ref={ref}
+          {...props}
+        >
+          {SideBarTrigger}
+          {children}
+        </div>
+      )
+    }
+
+    if (isMobile && isDefaultMobile) {
+      return (
+        <Sheet open={openMobile} onOpenChange={setOpenMobile} {...props}>
+          <SheetContent
+            data-sidebar="sidebar"
+            data-mobile="true"
+            className={cn(
+              "w-[--sidebar-width] bg-sidebar px-1 text-sidebar-foreground [&>button]:hidden",
+              className
+            )}
+            style={
+              {
+                "--sidebar-width": SIDEBAR_WIDTH_MOBILE,
+              } as React.CSSProperties
             }
-            return child
-          })}
-        </Flex>
-      </Box>
+            // @ts-expect-error TODO: fix error
+            side={side}
+          >
+            {SideBarTrigger}
+            <SheetTitle className="sr-only">Sidebar</SheetTitle>
+            <SheetDescription className="sr-only">Sidebar</SheetDescription>
+            <div className="flex h-full w-full flex-col">{children}</div>
+          </SheetContent>
+        </Sheet>
+      )
+    }
+
+    return (
+      <>
+        {isMobile && openMobile && (
+          <div
+            className="fixed inset-0 z-30 cursor-pointer bg-black bg-opacity-50 transition-opacity"
+            onClick={() => setOpenMobile(false)}
+            onKeyDown={(e) => {
+              if (e.key === "Enter" || e.key === " ") {
+                // toggleSidebar()
+              }
+            }}
+          />
+        )}
+
+        <div
+          ref={ref}
+          className={cn(
+            "group peer relative text-sidebar-foreground",
+            isMobile && openMobile
+              ? "fixed inset-y-0 left-0 z-40 w-[--sidebar-width]"
+              : "hidden",
+            !isMobile && "flex"
+          )}
+          data-state={state}
+          data-collapsible={state === "collapsed" ? collapsible : ""}
+          data-variant={variant}
+          data-side={side}
+          data-sidebar="sidebar"
+          data-mobile="true"
+        >
+          {SideBarTrigger}
+          {/* This is what handles the sidebar gap on desktop */}
+          <div
+            className={cn(
+              "relative h-svh w-[--sidebar-width] bg-transparent transition-[width] duration-200 ease-linear",
+              "group-data-[collapsible=offcanvas]:w-0",
+              "group-data-[side=right]:rotate-180",
+              variant === "floating" || variant === "inset"
+                ? "group-data-[collapsible=icon]:w-[calc(var(--sidebar-width-icon)_+_theme(spacing.4))]"
+                : "group-data-[collapsible=icon]:w-[--sidebar-width-icon]"
+            )}
+          />
+          <div
+            className={cn(
+              "fixed inset-y-0 z-10 flex h-svh w-[--sidebar-width] transition-[left,right,width] duration-200 ease-linear",
+              isMobile && openMobile ? "z-40 bg-sidebar shadow-lg" : "",
+              side === "left"
+                ? "left-0 group-data-[collapsible=offcanvas]:left-[calc(var(--sidebar-width)*-1)]"
+                : "right-0 group-data-[collapsible=offcanvas]:right-[calc(var(--sidebar-width)*-1)]",
+              // Adjust the padding for floating and inset variants.
+              variant === "floating" || variant === "inset"
+                ? "p-2 group-data-[collapsible=icon]:w-[calc(var(--sidebar-width-icon)_+_theme(spacing.4)_+2px)]"
+                : "group-data-[collapsible=icon]:w-[--sidebar-width-icon] ",
+              className
+            )}
+            {...props}
+          >
+            <div
+              data-sidebar="sidebar"
+              className={cn(
+                "flex h-full w-full flex-col bg-sidebar px-1 group-data-[variant=floating]:rounded-lg group-data-[variant=floating]:border group-data-[variant=floating]:border-sidebar-border group-data-[variant=floating]:shadow",
+                className
+              )}
+            >
+              {children}
+            </div>
+          </div>
+        </div>
+      </>
     )
+  }
+)
+Sidebar.displayName = "Sidebar"
 
-  const renderItem = () => (
-    <Flex
-      align="center"
-      className={`h-10 cursor-pointer rounded transition-all duration-300 ease-in-out ${
-        minified ? "w-10 px-2" : "mx-2 px-2"
-      } hover:bg-gray-7`}
-    >
-      <Box className="flex items-center">{renderIcon()}</Box>
-      <Box
-        className={`flex flex-1 items-center overflow-hidden transition-[width,opacity] duration-300 ease-in-out ${
-          minified ? "w-0 opacity-0" : "w-auto opacity-100"
-        }`}
+const SidebarTrigger = React.forwardRef<
+  React.ElementRef<typeof Button>,
+  React.ComponentProps<typeof Button>
+>(
+  (
+    {
+      className,
+      onClick,
+      icon,
+      isSecondary,
+      mobileSidebarTrigger,
+      srLabel,
+      ...props
+    },
+    ref
+  ) => {
+    const {
+      toggleSidebar,
+      toggleSecondarySidebar,
+      isMobile,
+      openMobile,
+      setOpenMobile,
+    } = useSidebar()
+
+    return (
+      <IconButton
+        ref={ref}
+        data-sidebar="trigger"
+        variant="soft"
+        className={cn("h-7 w-7", className)}
+        onClick={(event) => {
+          onClick?.(event)
+          if (isMobile) {
+            if (mobileSidebarTrigger) {
+              setOpenMobile(!openMobile)
+            }
+            if (isSecondary) {
+              toggleSecondarySidebar()
+            } else {
+              toggleSidebar()
+            }
+          } else {
+            if (isSecondary) {
+              toggleSecondarySidebar()
+            } else {
+              toggleSidebar()
+            }
+          }
+        }}
+        {...props}
       >
-        {renderLabel()}
-        {renderBadge()}
-        {renderExpandIcon()}
-      </Box>
-    </Flex>
-  )
+        {icon ? icon : <PanelLeft />}
+        <span className="sr-only">{srLabel ? srLabel : "Toggle Sidebar"}</span>
+      </IconButton>
+    )
+  }
+)
+SidebarTrigger.displayName = "SidebarTrigger"
 
-  const content = (
-    <Box>
-      {renderItem()}
-      {renderChildren()}
-    </Box>
-  )
+const SidebarRail = React.forwardRef<
+  HTMLButtonElement,
+  React.ComponentProps<"button">
+>(({ className, ...props }, ref) => {
+  const { toggleSidebar } = useSidebar()
 
   return (
-    <DropdownMenu.Root>
-      {children ? (
-        <DropdownMenu.Trigger
-          disabled={!minified}
-          onClick={minified ? undefined : toggleExpand}
-        >
-          {content}
-        </DropdownMenu.Trigger>
-      ) : to ? (
-        <RouterLink to={to}>{content}</RouterLink>
-      ) : (
-        <Box onClick={onClick}>{content}</Box>
+    <button
+      ref={ref}
+      data-sidebar="rail"
+      aria-label="Toggle Sidebar"
+      tabIndex={-1}
+      onClick={toggleSidebar}
+      title="Toggle Sidebar"
+      className={cn(
+        "-translate-x-1/2 group-data-[side=left]:-right-4 absolute inset-y-0 z-20 hidden w-4 transition-all ease-linear after:absolute after:inset-y-0 after:left-1/2 after:w-[2px] hover:after:bg-sidebar-border group-data-[side=right]:left-0 sm:flex",
+        "[[data-side=left]_&]:cursor-w-resize [[data-side=right]_&]:cursor-e-resize",
+        "[[data-side=left][data-state=collapsed]_&]:cursor-e-resize [[data-side=right][data-state=collapsed]_&]:cursor-w-resize",
+        "group-data-[collapsible=offcanvas]:translate-x-0 group-data-[collapsible=offcanvas]:hover:bg-sidebar group-data-[collapsible=offcanvas]:after:left-full",
+        "[[data-side=left][data-collapsible=offcanvas]_&]:-right-2",
+        "[[data-side=right][data-collapsible=offcanvas]_&]:-left-2",
+        className
       )}
-      <DropdownMenu.Content>
-        <Flex align="center" justify={"center"} className="px-4 py-2">
-          {renderIcon()}
-          <Text mx="2">{label}</Text>
-        </Flex>
-        <DropdownMenu.Separator />
-        {React.Children.map(children, (child) => {
-          if (React.isValidElement(child)) {
-            const { label, to, onClick } = child.props
-            return (
-              <DropdownMenu.Item>
-                {to ? (
-                  <RouterLink to={to} className="w-full">
-                    <Flex align="center" className="px-4 py-2">
-                      <Text>{label}</Text>
-                    </Flex>
-                  </RouterLink>
-                ) : (
-                  <Box onClick={onClick} className="w-full">
-                    <Flex align="center" className="px-4 py-2">
-                      <Text>{label}</Text>
-                    </Flex>
-                  </Box>
-                )}
-              </DropdownMenu.Item>
-            )
-          }
-          return child
-        })}
-      </DropdownMenu.Content>
-    </DropdownMenu.Root>
+      {...props}
+    />
   )
-}
+})
+SidebarRail.displayName = "SidebarRail"
 
-type SidebarSubItemBaseProps = {
-  label: string
-  onClick?: () => void
-}
+const SidebarInset = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"main">
+>(({ className, ...props }, ref) => {
+  return (
+    <main
+      ref={ref}
+      className={cn(
+        "relative flex min-h-svh flex-1 flex-col bg-background",
+        "peer-data-[variant=inset]:min-h-[calc(100svh-theme(spacing.4))] md:peer-data-[state=collapsed]:peer-data-[variant=inset]:ml-2 md:peer-data-[variant=inset]:m-2 md:peer-data-[variant=inset]:ml-0 md:peer-data-[variant=inset]:rounded-xl md:peer-data-[variant=inset]:shadow",
+        className
+      )}
+      {...props}
+    />
+  )
+})
+SidebarInset.displayName = "SidebarInset"
 
-type SidebarSubItemWithToProps = SidebarSubItemBaseProps & {
-  to: string
-  onClick?: never
-}
+const SidebarInput = React.forwardRef<
+  React.ElementRef<typeof Input>,
+  React.ComponentProps<typeof Input>
+>(({ className, ...props }, ref) => {
+  return (
+    <Input
+      ref={ref}
+      data-sidebar="input"
+      className={cn(
+        "h-8 w-full bg-background shadow-none focus-visible:ring-2 focus-visible:ring-sidebar-ring",
+        className
+      )}
+      {...props}
+    />
+  )
+})
+SidebarInput.displayName = "SidebarInput"
 
-type SidebarSubItemWithOnClickProps = SidebarSubItemBaseProps & {
-  to?: never
-  onClick: () => void
-}
+const SidebarHeader = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div">
+>(({ className, ...props }, ref) => {
+  return (
+    <div
+      ref={ref}
+      data-sidebar="header"
+      className={cn("flex flex-col gap-2 p-2", className)}
+      {...props}
+    />
+  )
+})
+SidebarHeader.displayName = "SidebarHeader"
 
-export type SidebarSubItemProps =
-  | SidebarSubItemWithToProps
-  | SidebarSubItemWithOnClickProps
+const SidebarFooter = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div">
+>(({ className, ...props }, ref) => {
+  return (
+    <div
+      ref={ref}
+      data-sidebar="footer"
+      className={cn("flex flex-col gap-2 p-2", className)}
+      {...props}
+    />
+  )
+})
+SidebarFooter.displayName = "SidebarFooter"
 
-export function SidebarSubItem({ label, to, onClick }: SidebarSubItemProps) {
-  const handleClick = (e: React.MouseEvent) => {
-    e.preventDefault()
-    e.stopPropagation()
-    if (onClick) {
-      onClick()
-    }
+const SidebarSeparator = React.forwardRef<
+  React.ElementRef<typeof Separator>,
+  React.ComponentProps<typeof Separator>
+>(({ className, ...props }, ref) => {
+  return (
+    <Separator
+      ref={ref}
+      data-sidebar="separator"
+      className={cn("mx-2 w-auto bg-sidebar-border", className)}
+      {...props}
+    />
+  )
+})
+SidebarSeparator.displayName = "SidebarSeparator"
+
+const SidebarContent = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div">
+>(({ className, ...props }, ref) => {
+  return (
+    <div
+      ref={ref}
+      data-sidebar="content"
+      className={cn(
+        "flex min-h-0 flex-1 flex-col gap-2 overflow-auto group-data-[collapsible=icon]:overflow-hidden",
+        className
+      )}
+      {...props}
+    />
+  )
+})
+SidebarContent.displayName = "SidebarContent"
+
+const SidebarGroup = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div">
+>(({ className, ...props }, ref) => {
+  return (
+    <div
+      ref={ref}
+      data-sidebar="group"
+      className={cn("relative flex w-full min-w-0 flex-col p-2", className)}
+      {...props}
+    />
+  )
+})
+SidebarGroup.displayName = "SidebarGroup"
+
+const SidebarGroupLabel = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div"> & { asChild?: boolean }
+>(({ className, asChild = false, ...props }, ref) => {
+  const Comp = asChild ? Slot : "div"
+
+  return (
+    <Comp
+      ref={ref}
+      data-sidebar="group-label"
+      className={cn(
+        "flex h-12 shrink-0 items-center gap-3 rounded-md px-2 font-medium text-lg text-sidebar-foreground/70 outline-none ring-sidebar-ring transition-[margin,opacity] duration-200 ease-linear focus-visible:ring-2",
+        "group-data-[collapsible=icon]:-mt-8 group-data-[collapsible=icon]:opacity-0",
+        className
+      )}
+      {...props}
+    />
+  )
+})
+SidebarGroupLabel.displayName = "SidebarGroupLabel"
+
+const SidebarHeaderLabel = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div"> & { asChild?: boolean }
+>(({ className, asChild = false, ...props }, ref) => {
+  const Comp = asChild ? Slot : "div"
+
+  return (
+    <Comp
+      ref={ref}
+      data-sidebar="group-label"
+      className={cn(
+        "mt-2 flex h-8 shrink-0 items-center gap-3 rounded-md px-2 font-medium text-lg text-sidebar-foreground/70 outline-none ring-sidebar-ring transition-[margin,opa] duration-200 ease-linear focus-visible:ring-2 [&>span:last-child]:text-2xl [&>svg]:size-4 [&>svg]:shrink-0",
+        "group-data-[collapsible=icon]:p-0",
+        className
+      )}
+      {...props}
+    />
+  )
+})
+SidebarHeaderLabel.displayName = "SidebarHeaderLabel"
+
+const SidebarGroupAction = React.forwardRef<
+  HTMLButtonElement,
+  React.ComponentProps<"button"> & { asChild?: boolean }
+>(({ className, asChild = false, ...props }, ref) => {
+  const Comp = asChild ? Slot : "button"
+
+  return (
+    <Comp
+      ref={ref}
+      data-sidebar="group-action"
+      className={cn(
+        "absolute top-3.5 right-3 flex aspect-square w-5 items-center justify-center rounded-md p-0 text-sidebar-foreground outline-none ring-sidebar-ring transition-transform hover:bg-sidebar-accent hover:text-sidebar-accent-foreground focus-visible:ring-2 [&>svg]:size-4 [&>svg]:shrink-0",
+        // Increases the hit area of the button on mobile.
+        "after:-inset-2 after:absolute after:md:hidden",
+        "group-data-[collapsible=icon]:hidden",
+        className
+      )}
+      {...props}
+    />
+  )
+})
+SidebarGroupAction.displayName = "SidebarGroupAction"
+
+const SidebarGroupContent = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div">
+>(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    data-sidebar="group-content"
+    className={cn("w-full text-sm", className)}
+    {...props}
+  />
+))
+SidebarGroupContent.displayName = "SidebarGroupContent"
+
+const SidebarMenu = React.forwardRef<
+  HTMLUListElement,
+  React.ComponentProps<"ul">
+>(({ className, ...props }, ref) => (
+  <ul
+    ref={ref}
+    data-sidebar="menu"
+    className={cn("flex w-full min-w-0 flex-col gap-1", className)}
+    {...props}
+  />
+))
+SidebarMenu.displayName = "SidebarMenu"
+
+const SidebarMenuItem = React.forwardRef<
+  HTMLLIElement,
+  React.ComponentProps<"li">
+>(({ className, ...props }, ref) => (
+  <li
+    ref={ref}
+    data-sidebar="menu-item"
+    className={cn("group/menu-item relative", className)}
+    {...props}
+  />
+))
+SidebarMenuItem.displayName = "SidebarMenuItem"
+
+const sidebarMenuButtonVariants = cva(
+  "peer/menu-button group-data-[collapsible=icon]:!size-8 group-data-[collapsible=icon]:!p-2 flex w-full items-center gap-2 overflow-hidden rounded-md p-2 text-left text-sm outline-none ring-sidebar-ring transition-[width,height,padding] hover:bg-sidebar-accent hover:text-sidebar-accent-foreground focus-visible:ring-2 active:bg-sidebar-accent active:text-sidebar-accent-foreground disabled:pointer-events-none disabled:opacity-50 group-has-[[data-sidebar=menu-action]]/menu-item:pr-8 aria-disabled:pointer-events-none aria-disabled:opacity-50 data-[active=true]:bg-sidebar-accent data-[active=true]:font-medium data-[active=true]:text-sidebar-accent-foreground data-[state=open]:hover:bg-sidebar-accent data-[state=open]:hover:text-sidebar-accent-foreground [&>span:last-child]:truncate [&>svg]:size-4 [&>svg]:shrink-0",
+  {
+    variants: {
+      variant: {
+        default: "hover:bg-sidebar-accent hover:text-sidebar-accent-foreground",
+        outline:
+          "bg-background shadow-[0_0_0_1px_hsl(var(--sidebar-border))] hover:bg-sidebar-accent hover:text-sidebar-accent-foreground hover:shadow-[0_0_0_1px_hsl(var(--sidebar-accent))]",
+      },
+      size: {
+        default: "h-12 text-sm",
+        sm: "h-7 text-xs",
+        lg: "group-data-[collapsible=icon]:!p-0 h-12 text-sm",
+      },
+    },
+    defaultVariants: {
+      variant: "default",
+      size: "default",
+    },
   }
+)
 
-  const content = (
-    <Flex align="center">
-      <Text>{label}</Text>
-    </Flex>
+const SidebarMenuButton = React.forwardRef<
+  HTMLButtonElement,
+  React.ComponentProps<"button"> & {
+    asChild?: boolean
+    isActive?: boolean
+    isSelected?: boolean
+    isSubMenuSelected?: boolean
+    tooltip?: string | React.ComponentProps<typeof TooltipContent>
+  } & VariantProps<typeof sidebarMenuButtonVariants>
+>(
+  (
+    {
+      asChild = false,
+      isActive = false,
+      variant = "default",
+      size = "default",
+      tooltip,
+      className,
+      isSelected,
+      isSubMenuSelected,
+      ...props
+    },
+    ref
+  ) => {
+    const Comp = asChild ? Slot : "button"
+    const { isMobile, state, open } = useSidebar()
+    const button = (
+      <Comp
+        ref={ref}
+        data-sidebar="menu-button"
+        data-size={size}
+        data-active={isActive}
+        className={cn(
+          sidebarMenuButtonVariants({ variant, size }),
+          "hover:bg-[hsl(var(--sidebar-accent))] dark:hover:bg-[hsl(var(--sidebar-accent))]",
+          `${isSelected && open && "relative rounded-tl-[0px] rounded-bl-[0px] border-l-0 bg-[hsl(var(--sidebar-accent))] font-[600] before:absolute before:top-0 before:bottom-0 before:left-0 before:w-[4px] before:rounded-tr-[4px] before:rounded-br-[4px] before:bg-[hsl(var(--sidebar-foreground))] before:content-['']"}`,
+          `${(isSelected || isSubMenuSelected) && !open && " bg-[hsl(var(--sidebar-accent))] text-[hsl(var(--sidebar-foreground))]"}`,
+          className
+        )}
+        {...props}
+      />
+    )
+
+    if (!tooltip) {
+      return button
+    }
+
+    if (typeof tooltip === "string") {
+      tooltip = {
+        children: tooltip,
+      }
+    }
+
+    return (
+      <Tooltip>
+        <TooltipTrigger asChild>{button}</TooltipTrigger>
+        <TooltipContent
+          side="right"
+          align="center"
+          hidden={state !== "collapsed" || isMobile}
+          {...tooltip}
+        />
+      </Tooltip>
+    )
+  }
+)
+SidebarMenuButton.displayName = "SidebarMenuButton"
+
+const SidebarMenuAction = React.forwardRef<
+  HTMLButtonElement,
+  React.ComponentProps<"button"> & {
+    asChild?: boolean
+    showOnHover?: boolean
+  }
+>(({ className, asChild = false, showOnHover = false, ...props }, ref) => {
+  const Comp = asChild ? Slot : "button"
+
+  return (
+    <Comp
+      ref={ref}
+      data-sidebar="menu-action"
+      className={cn(
+        "absolute top-1.5 right-1 flex aspect-square w-5 items-center justify-center rounded-md p-0 text-sidebar-foreground outline-none ring-sidebar-ring transition-transform hover:bg-sidebar-accent hover:text-sidebar-accent-foreground focus-visible:ring-2 peer-hover/menu-button:text-sidebar-accent-foreground [&>svg]:size-4 [&>svg]:shrink-0",
+        // Increases the hit area of the button on mobile.
+        "after:-inset-2 after:absolute after:md:hidden",
+        "peer-data-[size=sm]/menu-button:top-1",
+        "peer-data-[size=default]/menu-button:top-1.5",
+        "peer-data-[size=lg]/menu-button:top-2.5",
+        "group-data-[collapsible=icon]:hidden",
+        showOnHover &&
+          "group-focus-within/menu-item:opacity-100 group-hover/menu-item:opacity-100 data-[state=open]:opacity-100 peer-data-[active=true]/menu-button:text-sidebar-accent-foreground md:opacity-0",
+        className
+      )}
+      {...props}
+    />
   )
+})
+SidebarMenuAction.displayName = "SidebarMenuAction"
 
-  const className =
-    "block rounded py-2 text-gray-12 transition-all duration-300 ease-in-out px-4 hover:bg-gray-7"
+const SidebarMenuBadge = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div">
+>(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    data-sidebar="menu-badge"
+    className={cn(
+      "pointer-events-none absolute right-1 flex h-5 min-w-5 select-none items-center justify-center rounded-md px-1 font-medium text-sidebar-foreground text-xs tabular-nums",
+      "peer-hover/menu-button:text-sidebar-accent-foreground peer-data-[active=true]/menu-button:text-sidebar-accent-foreground",
+      "peer-data-[size=sm]/menu-button:top-1",
+      "peer-data-[size=default]/menu-button:top-1.5",
+      "peer-data-[size=lg]/menu-button:top-2.5",
+      "group-data-[collapsible=icon]:hidden",
+      className
+    )}
+    {...props}
+  />
+))
+SidebarMenuBadge.displayName = "SidebarMenuBadge"
 
-  return to ? (
-    <RouterLink to={to} className={className}>
-      {content}
-    </RouterLink>
-  ) : (
-    <Box onClick={handleClick} className={className}>
-      {content}
-    </Box>
+const SidebarMenuSkeleton = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<"div"> & {
+    showIcon?: boolean
+  }
+>(({ className, showIcon = false, ...props }, ref) => {
+  // Random width between 50 to 90%.
+  const width = React.useMemo(() => {
+    return `${Math.floor(Math.random() * 40) + 50}%`
+  }, [])
+
+  return (
+    <div
+      ref={ref}
+      data-sidebar="menu-skeleton"
+      className={cn("flex h-8 items-center gap-2 rounded-md px-2", className)}
+      {...props}
+    >
+      {showIcon && (
+        <Skeleton
+          className="size-4 rounded-md"
+          data-sidebar="menu-skeleton-icon"
+        />
+      )}
+      <Skeleton
+        className="h-4 max-w-[--skeleton-width] flex-1"
+        data-sidebar="menu-skeleton-text"
+        style={
+          {
+            "--skeleton-width": width,
+          } as React.CSSProperties
+        }
+      />
+    </div>
   )
+})
+SidebarMenuSkeleton.displayName = "SidebarMenuSkeleton"
+
+const SidebarMenuSub = React.forwardRef<
+  HTMLUListElement,
+  React.ComponentProps<"ul">
+>(({ className, ...props }, ref) => (
+  <ul
+    ref={ref}
+    data-sidebar="menu-sub"
+    className={cn(
+      "mx-3.5 flex min-w-0 translate-x-px flex-col gap-1 border-sidebar-border border-l px-2.5 py-0.5",
+      "group-data-[collapsible=icon]:hidden",
+      className
+    )}
+    {...props}
+  />
+))
+SidebarMenuSub.displayName = "SidebarMenuSub"
+
+const SidebarMenuSubItem = React.forwardRef<
+  HTMLLIElement,
+  React.ComponentProps<"li">
+>(({ ...props }, ref) => <li ref={ref} {...props} />)
+SidebarMenuSubItem.displayName = "SidebarMenuSubItem"
+
+const SidebarMenuSubButton = React.forwardRef<
+  HTMLAnchorElement,
+  React.ComponentProps<"a"> & {
+    asChild?: boolean
+    size?: "sm" | "md"
+    isActive?: boolean
+    isSelected?: boolean
+  }
+>(
+  (
+    { asChild = false, size = "md", isActive, isSelected, className, ...props },
+    ref
+  ) => {
+    const Comp = asChild ? Slot : "a"
+    const { open } = useSidebar()
+    return (
+      <Comp
+        ref={ref}
+        data-sidebar="menu-sub-button"
+        data-size={size}
+        data-active={isActive}
+        className={cn(
+          "-translate-x-px flex h-7 min-w-0 items-center gap-2 overflow-hidden rounded-md px-2 text-sidebar-foreground outline-none ring-sidebar-ring hover:bg-sidebar-accent hover:text-sidebar-accent-foreground focus-visible:ring-2 active:bg-sidebar-accent active:text-sidebar-accent-foreground disabled:pointer-events-none disabled:opacity-50 aria-disabled:pointer-events-none aria-disabled:opacity-50 [&>span:last-child]:truncate [&>svg]:size-4 [&>svg]:shrink-0 [&>svg]:text-sidebar-accent-foreground",
+          "data-[active=true]:bg-sidebar-accent data-[active=true]:text-sidebar-accent-foreground",
+          "hover:bg-[hsl(var(--sidebar-item-primary))] dark:hover:bg-[hsl(var(--sidebar-item-primary))]",
+          size === "sm" && "text-xs",
+          size === "md" && "text-sm",
+          isSelected &&
+            open &&
+            "relative rounded-tl-[0px] rounded-bl-[0px] border-l-0 bg-[hsl(var(--sidebar-item-primary))] font-[600] before:absolute before:top-0 before:bottom-0 before:left-0 before:w-[3px] before:rounded-tr-[4px] before:rounded-br-[4px] before:bg-[hsl(var(--sidebar-primary-foreground))] before:content-['']",
+          "group-data-[collapsible=icon]:hidden",
+          className
+        )}
+        {...props}
+      />
+    )
+  }
+)
+SidebarMenuSubButton.displayName = "SidebarMenuSubButton"
+
+export {
+  Sidebar,
+  SidebarContent,
+  SidebarFooter,
+  SidebarGroup,
+  SidebarGroupAction,
+  SidebarGroupContent,
+  SidebarGroupLabel,
+  SidebarHeader,
+  SidebarHeaderLabel,
+  SidebarInput,
+  SidebarInset,
+  SidebarMenu,
+  SidebarMenuAction,
+  SidebarMenuBadge,
+  SidebarMenuButton,
+  SidebarMenuItem,
+  SidebarMenuSkeleton,
+  SidebarMenuSub,
+  SidebarMenuSubButton,
+  SidebarMenuSubItem,
+  SidebarProvider,
+  SidebarRail,
+  SidebarSeparator,
+  SidebarTrigger,
+  useSidebar,
 }
